+++ packages/taro-cli/src/util/index.js
@@ -359,9 +359,18 @@ exports.urlJoin = function () {
 
 exports.resolveScriptPath = function (p) {
   let realPath = p
+  const taroEnv = process.env.TARO_ENV
   const SCRIPT_EXT = exports.JS_EXT.concat(exports.TS_EXT)
   for (let i = 0; i < SCRIPT_EXT.length; i++) {
     const item = SCRIPT_EXT[i]
+    if (taroEnv) {
+      if (fs.existsSync(`${p}.${taroEnv}${item}`)) {
+        return `${p}.${taroEnv}${item}`
+      }
+      if (fs.existsSync(`${p}${path.sep}index.${taroEnv}${item}`)) {
+        return `${p}${path.sep}index.${taroEnv}${item}`
+      }
+    }
     if (fs.existsSync(`${p}${item}`)) {
       return `${p}${item}`
     }