+++ packages/taro-components/types/Textarea.d.ts
@@ -1,7 +1,7 @@
 import { ComponentType } from 'react'
 import { StandardProps, CommonEventFunction, FormItemProps } from './common'
 
-interface TextareaProps extends StandardProps, FormItemProps {
+export interface TextareaProps extends StandardProps, FormItemProps {
 
   /**
    * 输入框的内容