+++ packages/taro-transformer-wx/src/index.ts
@@ -6,7 +6,21 @@ import * as ts from 'typescript'
 import { Transformer } from './class'
 import { setting, findFirstIdentifierFromMemberExpression, isContainJSXElement, codeFrameError, isArrayMapCallExpression, getSuperClassCode } from './utils'
 import * as t from 'babel-types'
-import { DEFAULT_Component_SET, INTERNAL_SAFE_GET, TARO_PACKAGE_NAME, REDUX_PACKAGE_NAME, MOBX_PACKAGE_NAME, IMAGE_COMPONENTS, INTERNAL_INLINE_STYLE, THIRD_PARTY_COMPONENTS, INTERNAL_GET_ORIGNAL, setLoopOriginal, GEL_ELEMENT_BY_ID, lessThanSignPlacehold } from './constant'
+import {
+  DEFAULT_Component_SET,
+  INTERNAL_SAFE_GET,
+  TARO_PACKAGE_NAME,
+  REDUX_PACKAGE_NAME,
+  MOBX_PACKAGE_NAME,
+  IMAGE_COMPONENTS,
+  INTERNAL_INLINE_STYLE,
+  THIRD_PARTY_COMPONENTS,
+  INTERNAL_GET_ORIGNAL,
+  setLoopOriginal,
+  GEL_ELEMENT_BY_ID,
+  lessThanSignPlacehold,
+  COMPONENTS_PACKAGE_NAME
+} from './constant'
 import { Adapters, setAdapter, Adapter } from './adapter'
 import { Options, setTransformOptions, buildBabelTransformOptions } from './options'
 import { get as safeGet } from 'lodash'
@@ -385,6 +399,16 @@ export default function transform (options: Options): TransformResult {
     },
     JSXOpeningElement (path) {
       const { name } = path.node.name as t.JSXIdentifier
+      const binding = path.scope.getBinding(name)
+      if (process.env.NODE_ENV !== 'test' && DEFAULT_Component_SET.has(name) && binding && binding.kind === 'module') {
+        const bindingPath = binding.path
+        if (bindingPath.parentPath.isImportDeclaration()) {
+          const source = bindingPath.parentPath.node.source
+          if (source.value !== COMPONENTS_PACKAGE_NAME) {
+            throw codeFrameError(bindingPath.parentPath.node, `内置组件名: '${name}' 只能从 ${COMPONENTS_PACKAGE_NAME} 引入。`)
+          }
+        }
+      }
       if (name === 'Provider') {
         const modules = path.scope.getAllBindings('module')
         const providerBinding = Object.values(modules).some((m: Binding) => m.identifier.name === 'Provider')
@@ -492,11 +516,11 @@ export default function transform (options: Options): TransformResult {
       path.traverse({
         ImportDefaultSpecifier (path) {
           const name = path.node.local.name
-          DEFAULT_Component_SET.has(name) || names.push(name)
+          names.push(name)
         },
         ImportSpecifier (path) {
           const name = path.node.imported.name
-          DEFAULT_Component_SET.has(name) || names.push(name)
+          names.push(name)
           if (source === TARO_PACKAGE_NAME && name === 'Component') {
             path.node.local = t.identifier('__BaseComponent')
           }