+++ packages/taro-weapp/src/native-api.js
@@ -2,7 +2,8 @@ import {
   onAndSyncApis,
   noPromiseApis,
   otherApis,
-  initPxTransform
+  initPxTransform,
+  Link
 } from '@tarojs/taro'
 import { cacheDataSet, cacheDataGet } from './data-cache'
 import { queryToJson, getUniqueKey } from './util'
@@ -34,6 +35,12 @@ const RequestQueue = {
     }
   }
 }
+
+function taroInterceptor (chain) {
+  return request(chain.requestParams)
+}
+
+const link = new Link(taroInterceptor)
 
 function request (options) {
   options = options || {}
@@ -195,7 +202,8 @@ function canIUseWebp () {
 
 export default function initNativeApi (taro) {
   processApis(taro)
-  taro.request = request
+  taro.request = link.request.bind(link)
+  taro.addInterceptor = link.addInterceptor.bind(link)
   taro.getCurrentPages = getCurrentPages
   taro.getApp = getApp
   taro.requirePlugin = requirePlugin